Alocacao dinamica

+++ Q01.c
@@ -3,9 +3,9 @@
 #include <math.h>
 
 
-long int func(long int *vetor, long int inicio, long int fim)
+int func(signed int *vetor, int inicio, int fim)
 {
-    long int mid = floor((fim + inicio)/2);
+    int mid = floor((fim + inicio)/2);
     if (inicio == fim)
         return inicio;
     if (vetor[inicio] >= vetor[inicio+1])
@@ -24,31 +24,31 @@ long int func(long int *vetor, long int inicio, long int fim)
 int main()
 {
     // Inicializa as variaveis
-    long int qtdTestes;
-    long int *tamanhos;
-    signed long int **vetores;
+    int qtdTestes;
+    int *tamanhos;
+    signed int **vetores;
 
     // Le as entradas
     setbuf(stdin, NULL);
-    scanf("%ld", &qtdTestes);
-    tamanhos = (long int*)malloc(qtdTestes * sizeof(long int));
-    vetores = (signed long int**)malloc(qtdTestes * sizeof(signed long int*));
+    scanf("%d", &qtdTestes);
+    tamanhos = (int*)malloc(qtdTestes * sizeof(int));
+    vetores = (signed int**)malloc(qtdTestes * sizeof(signed int*));
     for (int i = 0; i < qtdTestes; i++){
-        scanf("%ld", &tamanhos[i]);
-        vetores[i] = (signed long int*)malloc(tamanhos[i] * sizeof(signed long int));
+        scanf("%d", &tamanhos[i]);
+        vetores[i] = (signed int*)malloc(tamanhos[i] * sizeof(signed int));
         for (long int n = 0; n < tamanhos[i]; n++){
-            scanf("%ld", &vetores[i][n]);
+            scanf("%d", &vetores[i][n]);
         }
     }
 
     // Chama funcao recursiva para todos os casos
-    for (long int i = 0; i < qtdTestes; i++){
-        printf("%ld\n", func(vetores[i], 0, tamanhos[i] - 1) + 1);
+    for (int i = 0; i < qtdTestes; i++){
+        printf("%d\n", func(vetores[i], 0, tamanhos[i] - 1) + 1);
     }
 
     // Liberando memoria alocada
     free(tamanhos);
-    for (long int i = 0; i < qtdTestes; i++){
+    for (int i = 0; i < qtdTestes; i++){
         free(vetores[i]);
     }
     free(vetores);
